ai-analyst-backend/pyproject.toml
[project]
name = "ai-data-analyst-backend"
version = "0.1.0"
description = "LangGraph AI Data Analyst"
authors = [
    {name = "Sardar"}
]
requires-python = "^3.12"
dependencies = [
    "langgraph (>=1.1.3,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "sse-starlette (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
